refactor(serverTCP): :recycle: Multiplayer class
creating the Multiplayer game mode class.
This commit is just a checkpoint for refactoring the methods from the
previous merge commit. The code might not be working yet.
:construction: WIP: refactoring the Multiplayer game mode :construction:
